Fixed some server errors, added is followed in user data reponse

+++ server/database/followings.py
@@ -5,16 +5,16 @@ class Followings(Table):
     def follow(self, **kwargs):
         self._database.execute_and_commit('''
         INSERT INTO followings
-        (follower_id, followed_id)
+        (follower_id, user_id)
         VALUES
-        (%(follower_id)s, %(followed_id)s)
+        (%(follower_id)s, %(user_id)s)
         ''', kwargs)
     
     def unfollow(self, **kwargs):
         self._database.execute_and_commit('''
         DELETE FROM followings
         WHERE follower_id = %(follower_id)s
-        AND followed_id = %(followed_id)s
+        AND user_id = %(user_id)s
         ''', kwargs)
     
     def get_feed(self, **kwargs):
@@ -35,7 +35,7 @@ class Followings(Table):
             WHERE author_id = profiles.user_id
         ) FROM posts
         INNER JOIN followings
-        ON author_id = followed_id
+        ON author_id = user_id
         WHERE follower_id = %(follower_id)s
         ORDER BY id DESC
         LIMIT %(limit)s
@@ -46,18 +46,18 @@ class Followings(Table):
         return self._database.fetch_one('''
         SELECT count(*) AS total_count FROM posts
         INNER JOIN followings
-        ON author_id = followed_id
+        ON author_id = user_id
         WHERE follower_id = %(follower_id)s
         ''', kwargs)
 
     def get_followings(self, **kwargs):
         return self._database.fetch_all('''
-        SELECT followed_id, u.login, p.photo_url
+        SELECT user_id, u.login, p.photo_url
         FROM followings
         INNER JOIN users AS u
-        ON followed_id = u.id
+        ON user_id = u.id
         INNER JOIN profiles AS p
-        ON followed_id = p.user_id
+        ON user_id = p.user_id
         WHERE follower_id = %(follower_id)s
         ''', kwargs)
